changing

+++ Shape.cs
@@ -7,6 +7,6 @@ namespace OOpShapes
 {
     abstract class Shape
     {
-        public abstract int GetArea();
+        public abstract double GetArea();
     }
 }
